fix(db): traces.t_start_ms / duration_ms 를 BigInteger 로
증상: 통화 시도 시 STT trace 저장이 asyncpg DataError 로 실패.
'invalid input for query argument $5: 1778768004905 (value out of int32 range)'

원인: t_start_ms 컬럼이 INTEGER (int32, max 2.1B) 인데 epoch ms (1.7e12) 박혀
overflow. Trace 저장 실패 → Session rollback → 후속 build_runtime 도
PendingRollbackError → 응답 못 만듦 (= 사용자가 본 '응답 안 옴' 증상).

해결:
- models/trace.py: t_start_ms, duration_ms 를 Integer → BigInteger
- alembic 0002 revision: ALTER COLUMN (Postgres TYPE 변경, 데이터 보존)

duration_ms 도 안전 위해 BigInteger (현재 ms 라 32bit 충분하지만 일관성).

+++ backend/src/infrastructure/models/trace.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import Optional
 
-from sqlalchemy import JSON, ForeignKey, Integer, String, Text
+from sqlalchemy import JSON, BigInteger, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from ..db import Base
@@ -20,8 +20,9 @@ class Trace(Base):
     parent_id: Mapped[Optional[int]] = mapped_column(ForeignKey("traces.id"), nullable=True, index=True)
     name: Mapped[str] = mapped_column(String(255), nullable=False)
     kind: Mapped[str] = mapped_column(String(16), default="span")  # turn | llm | tool | tts | stt | span
-    t_start_ms: Mapped[int] = mapped_column(Integer, default=0)  # epoch ms
-    duration_ms: Mapped[int] = mapped_column(Integer, default=0)
+    # epoch ms (1.7e12+) — int32 overflow 방지 위해 BigInteger.
+    t_start_ms: Mapped[int] = mapped_column(BigInteger, default=0)
+    duration_ms: Mapped[int] = mapped_column(BigInteger, default=0)
     input_json: Mapped[dict] = mapped_column(JSON, default=dict)
     output_text: Mapped[str] = mapped_column(Text, default="")
     meta_json: Mapped[dict] = mapped_column(JSON, default=dict)
